chore: static type checking and typetests
