--- packages/react-native-reanimated/src/featureFlags/getStaticFlag.ts
@@ -1,7 +1,0 @@
-'use strict';
-
-import { ReanimatedModule } from '../ReanimatedModule';
-
-export function getStaticFeatureFlag(name: string): boolean {
-  return ReanimatedModule.getStaticFeatureFlag(name);
-}
