+++ packages/react-native-reanimated/src/mutables.ts
@@ -10,7 +10,7 @@ import {
 
 import { IS_JEST, logger, ReanimatedError, SHOULD_BE_USE_WEB } from './common';
 import type { Mutable } from './commonTypes';
-import { DynamicFlags } from './featureFlags/dynamicFlags';
+import { DynamicFlags } from './featureFlags';
 import { isFirstReactRender, isReactRendering } from './reactUtils';
 import { valueSetter } from './valueSetter';
