--- packages/react-native-worklets/src/featureFlags/dynamicFlags.ts
@@ -1,44 +1,0 @@
-'use strict';
-import { logger } from '../logger';
-import { WorkletsModule } from '../WorkletsModule';
-
-type DynamicFlagsType = {
-  EXAMPLE_DYNAMIC_FLAG: boolean;
-  init(): void;
-  setFlag(name: DynamicFlagName, value: boolean): void;
-};
-type DynamicFlagName = keyof Omit<Omit<DynamicFlagsType, 'setFlag'>, 'init'>;
-
-export const DynamicFlags: DynamicFlagsType = {
-  EXAMPLE_DYNAMIC_FLAG: true,
-
-  init() {
-    Object.keys(DynamicFlags).forEach((key) => {
-      if (key !== 'init' && key !== 'setFlag') {
-        WorkletsModule.setDynamicFeatureFlag(
-          key,
-          DynamicFlags[key as DynamicFlagName]
-        );
-      }
-    });
-  },
-  setFlag(name, value) {
-    if (name in DynamicFlags) {
-      DynamicFlags[name] = value;
-      WorkletsModule.setDynamicFeatureFlag(name, value);
-    } else {
-      logger.warn(
-        `The feature flag: '${name}' no longer exists, you can safely remove invocation of \`setDynamicFeatureFlag('${name}')\` from your code.`
-      );
-    }
-  },
-};
-DynamicFlags.init();
-
-// Public API function to update a feature flag
-export function setDynamicFeatureFlag(
-  name: DynamicFlagName,
-  value: boolean
-): void {
-  DynamicFlags.setFlag(name, value);
-}
