--- packages/react-native-worklets/src/featureFlags/getStaticFlag.ts
@@ -1,7 +1,0 @@
-'use strict';
-
-import { WorkletsModule } from '../WorkletsModule';
-
-export function getStaticFeatureFlag(name: string): boolean {
-  return WorkletsModule.getStaticFeatureFlag(name);
-}
